requirements.txt
datasets==2.6.1
donut==0.2.2
donut_python==1.0.8
fastapi==0.88.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
requests==2.28.1
streamlit==1.14.0
torch==1.13.0
transformers==4.25.0.dev0
uvicorn==0.20.0
